PLUS-trunk:  * Deny reqdel on a non-empty directory, decrement request counter when successful.    * TODO: Decrement the request counter of the user who created the request, not the user who issued the command [only a problem when user.isAdmin() is true]  * Fixed a possible NPE in replaceIdent() created for r1806  * Fixed a logic bug in IRCUsermanagement finduser command.  * Misc. non-functional changes. Nothing spectacular, mostly formatting / style edits which will make it easier to share code between my public/private repositories.
Thanks to _Dink_ and N0S on IRC for the nudge.

Closes: #173

git-svn-id: https://drftpd.svn.sourceforge.net/svnroot/drftpd/plus/2.0/trunk@1820 c83153ee-f3e8-0310-b165-dc3a72e09794

+++ src/net/drmods/plugins/irc/Request.java
@@ -259,7 +259,7 @@ public class Request extends IRCCommand {
             dir.createDirectory("REQUEST-by." + requser + "-" + dirName);
             LinkedRemoteFileInterface reqdir = dir.getFile("REQUEST-by." + requser + "-" + dirName);
             reqdir.setOwner(requser);
-            user.getKeyedMap().setObject(Request.REQUESTS, user.getKeyedMap().getObjectInt(Request.REQUESTS)+1);;
+            user.getKeyedMap().setObject(Request.REQUESTS, user.getKeyedMap().getObjectInt(Request.REQUESTS)+1);
             ArrayList<String> forceToChannels = new ArrayList<String>();
             forceToChannels.add(msgc.getDest());
             getGlobalContext().dispatchFtpEvent(new DirectorySiteBotEvent(
@@ -318,13 +318,17 @@ public class Request extends IRCCommand {
                 if (file.isDirectory()) {
                     if (file.getName().endsWith(dirName)) {
                         nodir = false;
-                        if (file.getUsername().equals(user.getName()) || user.isAdmin()) {
+                        if (file.dirSize() > 0 && (file.getUsername().equals(user.getName()) || user.isAdmin())) {
                             file.delete();
                             deldir = true;
                             ArrayList<String> forceToChannels = new ArrayList<String>();
                             forceToChannels.add(msgc.getDest());
-                            getGlobalContext().dispatchFtpEvent(new DirectorySiteBotEvent(   
-                            		user, "REQDEL", file, forceToChannels));   
+                            getGlobalContext().dispatchFtpEvent(new DirectorySiteBotEvent(
+                            		user, "REQDEL", file, forceToChannels));
+                            // decrement request counter
+                            int reqsMade = user.getKeyedMap().getObjectInt(Request.REQUESTS);
+                            if (reqsMade > 0)
+                            	user.getKeyedMap().setObject(Request.REQUESTS, --reqsMade);
                             break;
                         } else {
                             out.add(ReplacerUtils.jprintf("reqdel.notowner", env, Request.class));
